requirements.txt
alembic==1.2.1
astroid==2.2.5
autopep8==1.4.4
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.3.9
cffi==1.13.1
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.8
firebase-admin==3.1.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.1
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.5.0
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.3
httplib2==0.14.0
idna==2.8
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.3
jwt==0.6.1
lazy-object-proxy==1.4.0
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
mysqlclient==1.4.4
mysql-connector
peewee==3.11.2
pipenv==2018.11.26
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.3.5
uritemplate==3.0.0
urllib3==1.25.6
virtualenv==16.5.0
virtualenv-clone==0.5.3
Werkzeug==0.16.0
wrapt==1.11.1
